Colour output as if gtest had done it
Colour output on Windows is lost when piping. This is fine on systems with
terminals that understand ANSI escape sequences as it is just data a terminal
understands when it encounters it. This is not the case on Windows. The stream
itself must be modified to output colour but if it is not a console, this
modification fails (gtest silently ignores the error).

So for the sake of consistency of test output across the platforms, colours are
disabled when running the tests and then set explicitly as if it had been
enabled.