Run a full test after passing a failing test in a distinct test session
It is unclear what it happening when a failing test that is fixed causes a full
test run to be done. The session indicator now separates the filtered run from
the full run.

Also...

Return output from a streamed_call to aid debugging and be able to provide a
verbose mode.

Add another abstraction layer to the underlying file watching functionality so
that a drop-in replacement using file event notification is possible.